requirements.txt
# Resume Relevance Checker - Innomatics Research Labs
# Complete AI-Powered Resume Analysis System
# Supports Hugging Face, spaCy, and top AI models

# Core Web Framework Dependencies
streamlit>=1.28.1
plotly>=5.17.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and ORM
sqlalchemy>=2.0.23
alembic>=1.13.1

# Document Processing and Text Extraction
PyPDF2>=3.0.1
python-docx>=1.1.0
aiofiles>=23.2.1

# Core Machine Learning and Data Processing
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.11.4

# String Matching and Text Processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
nltk>=3.8.1

# Hugging Face Ecosystem - Complete Suite
transformers>=4.36.0
sentence-transformers>=2.2.2
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
huggingface-hub>=0.19.4
datasets>=2.14.0
accelerate>=0.24.0
tokenizers>=0.15.0
safetensors>=0.4.3

# spaCy NLP Processing
spacy>=3.7.2
# Note: Run 'python -m spacy download en_core_web_sm' after installation

# Additional Hugging Face Models and Tools
evaluate>=0.4.0
peft>=0.6.0
bitsandbytes>=0.41.0

# LangChain for Enhanced AI Capabilities
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# HTTP Requests and API Integration
requests>=2.31.0
httpx>=0.24.0

# Environment and Configuration Management
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development and Testing Tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.0.0
flake8>=6.0.0

# Additional Utilities and Support
colorama>=0.4.6
tqdm>=4.66.0
click>=8.1.0
typing-extensions>=4.8.0

# Security and Authentication
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring and Logging
loguru>=0.7.0
prometheus-client>=0.17.0

# Tunneling and Public Access
pyngrok>=7.0.0

# Optional: For advanced model fine-tuning
# apex>=0.1  # Uncomment if using NVIDIA GPUs
# flash-attn>=2.0.0  # Uncomment for flash attention

# Production deployment
gunicorn>=21.2.0
psycopg2-binary>=2.9.0  # For PostgreSQL
redis>=5.0.0  # For caching